% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{bubble}
\alias{bubble}
\title{Perform hierarchical clustering using bubble clustering methods.  Produces an hclust object}
\usage{
bubble(data, level = 3, Brad = 1, verbose = FALSE)
}
\arguments{
\item{data}{A numerical matrix}

\item{level}{Level of desired bubble coverage generates n*10^level bubbles are generated}

\item{Brad}{maximum bubble radius in the range of 0-1 of maximum distance between points}

\item{verbose}{Boolean indicator of progress bar presence or absence}
}
\value{
hclust object of hierarcical tree of \code{data}
}
\description{
Perform hierarchical clustering using bubble clustering methods.  Produces an hclust object
}
\examples{
bubble(myData, 4, .75)
bubble(myData)
}
