% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble.R
\name{HClustStabHist}
\alias{HClustStabHist}
\title{Calculate the snip-rebuild distance using hclust() for each taxa and returns the value as a vector}
\usage{
HClustStabHist(data, method = "complete")
}
\arguments{
\item{data}{A numerical matrix}

\item{method}{Desired methodology for joining clusters in hclust() and as such should be  one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}
}
\value{
Returns the snip-rebuild distance for each sample in \code{data} using hclust() as a vector
}
\description{
Calculate the snip-rebuild distance using hclust() for each taxa and returns the value as a vector
}
\examples{
BubStab(myData, "complete")
BubStab(myData, method= "ward.D")

}
